# Using python-Levenshtein for similarity between lemmas, for performance