@gutfeeling 2020-09-17
@CubieDev Since you are working on the package quite a bit, I want to tell you one of my ideas.

As written in the README of this package, lemmatizers often have difficulties producing the same lemma for related words. For example, "operations" is reduced to "operation" and "operate" is reduced to "operate". 

Since this package knows that "operation" and "operate" are related, we could perhaps make a better lemmatizer?