@CubieDev 2020-09-17
@gutfeeling 
> @CubieDev Since you are working on the package quite a bit, I want to tell you one of my ideas.
> 
> As written in the README of this package, lemmatizers often have difficulties producing the same lemma for related words. For example, "operations" is reduced to "operation" and "operate" is reduced to "operate".
> 
> Since this package knows that "operation" and "operate" are related, we could perhaps make a better lemmatizer?

That sounds very interesting. I don't have much experience with how lemmatizers function, but that hasn't stopped me before. I feel like a good way to go about that would be to create a very large testing suite, and use it to test many of the existing lemmatizers. We could then find where existing lemmatizers fail, to see if they indeed fail with `operations`. Then we could use test-driven development to try and make the tests pass properly.