@gutfeeling 2020-09-17
@CubieDev 

Given what I saw in your work on this package, I think nothing can stop you once you put your mind to it :-D

Your approach of test-driven development to make a better lemmatizer out of this package sounds promising. I wanted to mention the following.

1. Lemmatizers convert a word to a base form that is present in the English dictionary. The definition of a lemmatizer, at least, doesn't tell you which base form it is going to reduce the word to. 
2. Therefore, a test would probably only check that `operations` and `operate` are not being reduced to the same base form, even though we know that they are related words.
3. In the case of `word_forms` the task would be to decide how to select the base form out of all related forms of a word. Maybe just selecting the first noun in alphabetic order (i.e the one that appears first in the dictionary) is an idea. If no nouns are present, we choose verbs. If no verbs, then adverbs, and so on. However, this might not be the most intuitive base form for the end user.
4. It would also be interesting to find out whether there is added value to creating such a lemmatizer.  Intuitively, I think there is. But would be nice to have some core NLP people comment on that. Maybe @chrisjbryant has an idea? His profile indicates that he is a researcher in NLP.