@chrisjbryant 2020-09-18
> @CubieDev
>
> 4. It would also be interesting to find out whether there is added value to creating such a lemmatizer.  Intuitively, I think there is. But would be nice to have some core NLP people comment on that. Maybe @chrisjbryant has an idea? His profile indicates that he is a researcher in NLP.

I do indeed work in NLP, but my intuition is that there are actually probably already enough public lemmatizers (I'd currently recommend [LemmInflect](https://github.com/bjascob/LemmInflect)) and what you're really describing for the "operation" vs "operate" case is a stemmer.

A bit of linguistics:
Inflectional morphology describes affixes that generally do not change the part-of-speech of a word. E.g. operation vs operations are both nouns, while operate, operates, operated, and operating, are all verbs.
Derivational morphology describes affixes that often (but not always) change the part-of-speech of a word. E.g. operate (verb) vs. operation (noun), quick (adj) vs. quickly (adv), sign (noun/verb) vs. resign (verb). 

From what I know, lemmatizers are generally only designed to handle inflectional morphology. While it certainly makes sense that "operate" and "operations" should return the same form, things get a bit more complicated if we also throw "cooperate" into the mix, which likewise comes from the same root. I'm guessing you probably don't want it to match "operate" though. 

This is where stemmers come in. Stemmers attempt to cut off all the affixes until you're left with the base stem. In the case of "operate" and "operations", this will probably be "operat". Note that the stem doesn't have to be a real word. Depending on the stemmer, "cooperate" might also get reduced to "operat" or possibly "cooperat". 

I'm telling you all of this just to give you a bit of background on the difference between lemmatizers and stemmers. Lemmatizers are far more common than stemmers today (although I do use a stemmer in one of my own repos), but I don't think the algorithms for either of them have changed much in the past 20 years. 

Consequently, I'd suggest taking a look at the different options that are available before deciding whether to develop your own. It might be better just to add some existing tools to this package to get the functionality you want!