@CubieDev 2020-09-18
@gutfeeling @chrisjbryant 
> (I'd currently recommend LemmInflect)
I googled a bunch yesterday, and came across them, sporting some pretty impressive accuracy percentages. Obviously, accuracy isn't everything, but their performance seems hard to beat. If I remember correctly, they use a set of "certain" rules first, and if a word does not fall within these rules, they let a machine learned model decide the suffix. It seems quite cool.

Regardless, I agree that we should look at different existing options before wanting to develop something ourselves. This is why I was googling, in fact. I'd rather not build something for dozens of hours, only to find it does the same thing as another project, but worse.