@gutfeeling 2020-09-18
@chrisjbryant  Thanks so much for your insight. It's invaluable!

@CubieDev Thanks for researching this on Google. I went ahead and implemented a really simple lemmatizer here https://github.com/gutfeeling/word_forms/blob/master/word_forms/lemmatizer.py

Feel free to check it out!