@CubieDev 2020-09-21
@gutfeeling I had some time to play around with the lemmatizer. It does seem to pretty consistently group words with the same base together. 
For example, `am`, `was`, `are`, `be`, `being`, `is`, `been`, etc. all result in `am`. I suppose this is to be expected, as all of these words are grouped together in `get_word_forms` nowadays.

When messing around I did find that `run` lemmatizes to `ran`. 
This is because of the center line in
https://github.com/gutfeeling/word_forms/blob/9f24160c6e0ec0fdf3674cdd16f1ca68cb8bc320/word_forms/lemmatizer.py#L8-L10
which sorts alphabetically, putting `ran` before `run`. There are probably more cases where the alphabetical sorting messes things up. Whether this is a big issue is another matter, as I suspect the main use case of a lemmatizer like this one is detecting whether two words are from the same base. 

As in, checking whether `lemmatize("runniness") == lemmatize("runny")`. In this case it doesn't matter much whether the lemmatizer returns `ran` or `run`.

I don't really have any suggestions on optimizations or improvements here, apart from unimportant changes like turning the list comprehension of line 8 into a set comprehension, so the sorting might *technically* be a bit faster, as the set is smaller than the list when there are duplicate elements. (E.g. when "running" is both a verb and an adjective)
This is about as inconsequential as a change could be, as the sorting is on a maximum of like 12 elements anyways, and the call to `get_word_forms` is the only real impact on the execution time.

The lemmatizer is cool, but is probably considerably less consistent than preexisting ones sadly.